Move IndexSearcher to internal package
